[ALL] Converted variables with expected bit-age to uint32/64_t and uintptr_t.
[GAFF]
Made the FNV "wrapper" functions just plain old overloads and removed the default argument for the versions that take an init value.
Added custom assert handler API.

[SHIBBOLETH]
Fixed rawRequestInterface() interfaces to take a Gaff::ReflectionHash instead of an unsigned int.

TODO: Convert assert() calls to GAFF_ASSERT()/GAFF_ASSERT_MSG() calls.

+++ src/Components/Shibboleth_CameraComponent.cpp
@@ -107,7 +107,7 @@ bool CameraComponent::save(Gaff::JSON& json)
 	return true;
 }
 
-void CameraComponent::updateFrustum(Object* object, unsigned long long)
+void CameraComponent::updateFrustum(Object* object, uint64_t)
 {
 	_frustum = _unstransformed_frustum;
 	_frustum.transform(object->getWorldTransform());
