+++ src/Game/Shibboleth_ShaderProgramLoader.cpp
@@ -45,7 +45,7 @@ ShaderProgramLoader::~ShaderProgramLoader(void)
 {
 }
 
-Gaff::IVirtualDestructor* ShaderProgramLoader::load(const char* file_name, unsigned long long, HashMap<AString, IFile*>& file_map)
+Gaff::IVirtualDestructor* ShaderProgramLoader::load(const char* file_name, uint64_t, HashMap<AString, IFile*>& file_map)
 {
 	assert(file_map.hasElementWithKey(AString(file_name)));
